Re-organized data for publishing
Put the essential efficient greedy solver C files into the root directory.
Removed testing and organizational files.
Added README.md.
Added comments in files.
