--- greedy_003/difference.h
@@ -1,4 +1,0 @@
-#include <stdlib.h>
-#include "problem.h"
-
-void difference(struct set left, struct set right, struct set* differenceSet);
