+++ greedy_003/dynamic_array_element.c
@@ -2,6 +2,7 @@
 
 #include "dynamic_array_element.h"
 
+/* dynamically sized array; doubles size if added element exceeds capacity; elements are of type `element' */
 void _double_capacity_element(struct dynamic_array_element* array) {
 	struct element* extended_data = calloc(2 * array->capacity, sizeof(struct element));		
 	for (int i = 0; i < array->capacity; ++i)
