+++ greedy_003/dynamic_array_element.h
