+++ greedy_003/dynamic_array_index_pair.h
