--- greedy_003/greedy_with_LUT/dynamically_sized_array.spec
@@ -1,8 +1,0 @@
-S P E C I F I C A T I O N
-=========================
-A dynamically sized array should..
-
-1.  have an initial size of 16.
-2.  support adding functionality.
-3.  resize to 32 if the 17-th element is added.
-4.  resize to n * 2 if the (n + 1)-th element is added.
