+++ greedy_003/greedy_with_LUT_core.c
