--- greedy_003/harmonic_number.h
@@ -1,1 +1,0 @@
-int harmonic_number(int n, double* Hn);
