--- greedy_003/problem.h
@@ -1,11 +1,0 @@
-struct set {
-	double cost;
-	int numberOfElements;
-	int* elements;
-};
-
-struct problem {
-	int numberOfSets;
-	int numberOfElements;
-	struct set* sets;
-};
