--- greedy_003/read.h
@@ -1,4 +1,0 @@
-#include "problem.h"
-
-int* copyIntegerArray(int source[], int numberOfElements);
-void readProblemDescription(char* path, struct problem* specificProblem);
