--- greedy_003/setcover_bruteforce.h
@@ -1,4 +1,0 @@
-#include "solution.h"
-#include "union.h"
-
-void setcover_bruteforce(struct problem* specificProblem, struct solution* specificSolution);
