--- greedy_003/solution.h
@@ -1,6 +1,0 @@
-struct solution {
-	int numberOfSets;
-	int numberOfSetsPicked;
-	int* pickedSets;
-	double cost;
-};
