--- greedy_003/test_client_greedy.bat
@@ -1,2 +1,0 @@
-@echo off
-for /F %%G in ('dir .\data /b') do test_client_greedy.exe .\data\%%G
