--- greedy_003/tests/200412_performance_listing_without_header
@@ -1,65 +1,0 @@
-6,0.000000e+000,0.000000e+000
-9,0.000000e+000,0.000000e+000
-15,0.000000e+000,0.000000e+000
-25,0.000000e+000,0.000000e+000
-27,0.000000e+000,0.000000e+000
-45,0.000000e+000,0.000000e+000
-81,0.000000e+000,0.000000e+000
-135,6.300000e-002,1.146531e-009
-157,0.000000e+000,0.000000e+000
-192,0.000000e+000,0.000000e+000
-243,6.530000e-001,1.128315e-009
-330,3.100000e-002,2.782647e-010
-405,3.427000e+000,7.661590e-010
-448,4.100000e-002,3.039895e-010
-450,2.786000e+000,7.717521e-010
-450,2.264000e+000,6.277863e-010
-450,2.527000e+000,6.998493e-010
-450,2.597000e+000,7.207312e-010
-450,2.591000e+000,7.158477e-010
-495,9.200000e-002,1.834254e-010
-500,0.000000e+000,0.000000e+000
-500,0.000000e+000,0.000000e+000
-500,0.000000e+000,0.000000e+000
-500,8.000000e-003,6.400000e-011
-500,0.000000e+000,0.000000e+000
-595,6.842000e+000,6.940180e-010
-595,6.806000e+000,6.882901e-010
-595,6.317000e+000,6.405576e-010
-595,7.023000e+000,7.048851e-010
-595,6.159000e+000,6.248501e-010
-715,2.860000e-001,1.366155e-010
-729,3.463700e+001,7.368474e-010
-760,1.408300e+001,5.908908e-010
-760,1.497400e+001,6.274995e-010
-760,1.675100e+001,7.057072e-010
-760,1.597600e+001,6.648066e-010
-760,1.346300e+001,5.600451e-010
-945,6.455300e+001,1.233041e-009
-945,4.615600e+001,8.823128e-010
-945,3.419000e+001,6.573200e-010
-945,3.763700e+001,7.198332e-010
-945,3.201200e+001,6.056618e-010
-1000,3.200000e-002,3.200000e-011
-1000,3.100000e-002,3.100000e-011
-1000,2.100000e-002,2.100000e-011
-1000,2.200000e-002,2.200000e-011
-1000,2.600000e-002,2.600000e-011
-1000,1.500000e-002,1.500000e-011
-1000,3.100000e-002,3.100000e-011
-1000,3.100000e-002,3.100000e-011
-1000,3.100000e-002,3.100000e-011
-1000,3.800000e-002,3.800000e-011
-1000,3.100000e-002,3.100000e-011
-1000,3.800000e-002,3.800000e-011
-1000,1.600000e-002,1.600000e-011
-1000,3.100000e-002,3.100000e-011
-1000,2.200000e-002,2.200000e-011
-1024,5.850000e-001,3.113278e-010
-1150,1.389000e+002,1.311674e-009
-1150,1.104540e+002,1.033000e-009
-1150,1.155380e+002,1.077655e-009
-1150,7.648100e+001,7.205590e-010
-1150,7.241800e+001,6.841808e-010
-1165,0.000000e+000,0.000000e+000
-1215,4.849370e+002,1.336254e-009
