--- greedy_003/tests/200420_performance_listing
@@ -1,70 +1,0 @@
-Number of sets | Elapsed time [sec] | Constant |
----------------+--------------------------------
-             6 |       0.000000e+000 |       0.000000e+000
-             9 |       0.000000e+000 |       0.000000e+000
-            15 |       0.000000e+000 |       0.000000e+000
-            25 |       0.000000e+000 |       0.000000e+000
-            27 |       0.000000e+000 |       0.000000e+000
-            45 |       1.000000e-003 |       1.496446e-009
-            81 |       8.000000e-003 |       1.129006e-009
-           135 |       6.300000e-002 |       1.146531e-009
-           157 |       0.000000e+000 |       0.000000e+000
-           192 |       3.000000e-003 |       3.390842e-010
-           243 |       6.440000e-001 |       1.112764e-009
-           330 |       3.000000e-002 |       2.692885e-010
-           405 |       3.432000e+000 |       7.672768e-010
-           448 |       4.300000e-002 |       3.188182e-010
-           450 |       2.805000e+000 |       7.770153e-010
-           450 |       2.336000e+000 |       6.477513e-010
-           450 |       2.533000e+000 |       7.015110e-010
-           450 |       2.606000e+000 |       7.232289e-010
-           450 |       2.591000e+000 |       7.158477e-010
-           495 |       9.200000e-002 |       1.834254e-010
-           500 |       0.000000e+000 |       0.000000e+000
-           500 |       0.000000e+000 |       0.000000e+000
-           500 |       0.000000e+000 |       0.000000e+000
-           500 |       0.000000e+000 |       0.000000e+000
-           500 |       0.000000e+000 |       0.000000e+000
-           595 |       6.874000e+000 |       6.972639e-010
-           595 |       6.873000e+000 |       6.950658e-010
-           595 |       6.703000e+000 |       6.796988e-010
-           595 |       7.128000e+000 |       7.154238e-010
-           595 |       6.378000e+000 |       6.470683e-010
-           715 |       2.920000e-001 |       1.394816e-010
-           729 |       3.521200e+001 |       7.490796e-010
-           760 |       1.417300e+001 |       5.946670e-010
-           760 |       1.508000e+001 |       6.319416e-010
-           760 |       1.690400e+001 |       7.121530e-010
-           760 |       1.609400e+001 |       6.697169e-010
-           760 |       1.355600e+001 |       5.639138e-010
-           945 |       3.583200e+001 |       6.844346e-010
-           945 |       3.489200e+001 |       6.669915e-010
-           945 |       3.449500e+001 |       6.631838e-010
-           945 |       3.797100e+001 |       7.262212e-010
-           945 |       3.228800e+001 |       6.108837e-010
-          1000 |       3.400000e-002 |       8.500000e-010
-          1000 |       3.000000e-002 |       7.500000e-010
-          1000 |       2.700000e-002 |       6.750000e-010
-          1000 |       2.400000e-002 |       6.000000e-010
-          1000 |       2.300000e-002 |       5.750000e-010
-          1000 |       2.500000e-002 |       6.250000e-010
-          1000 |       3.300000e-002 |       8.250000e-010
-          1000 |       3.300000e-002 |       8.250000e-010
-          1000 |       3.100000e-002 |       7.750000e-010
-          1000 |       3.200000e-002 |       8.000000e-010
-          1000 |       3.100000e-002 |       7.750000e-010
-          1000 |       2.800000e-002 |       7.000000e-010
-          1000 |       2.400000e-002 |       6.000000e-010
-          1000 |       3.300000e-002 |       8.250000e-010
-          1000 |       2.700000e-002 |       6.750000e-010
-          1024 |       6.040000e-001 |       3.214393e-010
-          1150 |       8.811000e+001 |       8.320489e-010
-          1150 |       6.957700e+001 |       6.507057e-010
-          1150 |       7.375200e+001 |       6.879053e-010
-          1150 |       7.211200e+001 |       6.793968e-010
-          1150 |       7.340900e+001 |       6.935435e-010
-          1165 |       1.200000e-002 |       7.397608e-010
-          1215 |       4.091580e+002 |       1.127443e-009
-          1272 |       1.014590e+002 |       6.649176e-010
-          1272 |       1.000090e+002 |       6.559783e-010
-          1272 |       8.967100e+001 |       5.877826e-010
