--- greedy_003/tests/tests_performance_greedy_with_LUT.out.csv
@@ -1,1 +1,0 @@
-6,0.000000e+000,0.000000e+0009,0.000000e+000,0.000000e+00015,0.000000e+000,0.000000e+00025,0.000000e+000,0.000000e+00027,0.000000e+000,0.000000e+00045,0.000000e+000,0.000000e+00081,0.000000e+000,0.000000e+000135,0.000000e+000,0.000000e+000157,0.000000e+000,0.000000e+000192,0.000000e+000,0.000000e+000243,0.000000e+000,0.000000e+000330,0.000000e+000,0.000000e+000405,1.600000e-002,1.448705e-009448,1.500000e-002,4.982462e-008450,1.600000e-002,1.994478e-009450,1.500000e-002,1.871713e-009450,1.500000e-002,1.869403e-009450,1.600000e-002,1.998177e-009450,1.500000e-002,1.864906e-009495,2.000000e-002,1.973817e-008500,0.000000e+000,0.000000e+000500,0.000000e+000,0.000000e+000500,0.000000e+000,0.000000e+000500,0.000000e+000,0.000000e+000500,0.000000e+000,0.000000e+000595,3.100000e-002,1.870968e-009595,1.600000e-002,9.627567e-010595,1.600000e-002,9.653488e-010595,2.200000e-002,1.313818e-009595,1.600000e-002,9.658342e-010715,8.500000e-002,2.903080e-008729,9.200000e-002,1.426766e-009760,4.100000e-002,1.307403e-009760,3.900000e-002,1.242092e-009760,3.000000e-002,9.605471e-010760,3.100000e-002,9.803984e-010760,3.300000e-002,1.043299e-009945,4.100000e-002,7.400765e-010945,5.200000e-002,9.393546e-010945,5.200000e-002,9.447413e-010945,5.100000e-002,9.217622e-010945,5.100000e-002,9.118415e-0101000,2.000000e-003,1.000000e-0081000,0.000000e+000,0.000000e+0001000,8.000000e-003,4.000000e-0081000,0.000000e+000,0.000000e+0001000,9.000000e-003,4.500000e-0081000,0.000000e+000,0.000000e+0001000,1.000000e-002,5.000000e-0081000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001000,0.000000e+000,0.000000e+0001024,0.000000e+000,0.000000e+0001150,7.100000e-002,7.710452e-0101150,7.100000e-002,7.636162e-0101150,6.900000e-002,7.401194e-0101150,6.900000e-002,7.475890e-0101150,6.900000e-002,7.496720e-0101165,0.000000e+000,0.000000e+0001215,2.690000e-001,9.006007e-0101272,8.300000e-002,6.918986e-0101272,8.500000e-002,7.091800e-0101272,6.900000e-002,5.753087e-0101272,7.800000e-002,6.507838e-0101272,8.500000e-002,7.099334e-0101400,8.400000e-002,5.484411e-0101400,1.010000e-001,6.595677e-0101400,1.000000e-001,6.491264e-0101400,1.010000e-001,6.582317e-0101400,1.000000e-001,6.512689e-0101534,1.180000e-001,6.078233e-0101534,1.120000e-001,5.787096e-0101534,1.130000e-001,5.842517e-0101534,1.120000e-001,5.748458e-0101534,1.120000e-001,5.795410e-0102000,8.000000e-003,2.000000e-0082000,0.000000e+000,0.000000e+0002000,1.000000e-002,2.500000e-0082000,0.000000e+000,0.000000e+0002000,1.600000e-002,4.000000e-0082000,0.000000e+000,0.000000e+0002000,0.000000e+000,0.000000e+0002000,0.000000e+000,0.000000e+0002000,0.000000e+000,0.000000e+0002000,1.700000e-002,4.250000e-0082187,9.210000e-001,5.285221e-0102241,0.000000e+000,0.000000e+0002304,1.800000e-002,1.695421e-0093000,8.000000e-003,8.888889e-0093000,1.200000e-002,1.333333e-0083000,0.000000e+000,0.000000e+0003000,0.000000e+000,0.000000e+0003000,2.000000e-002,2.222222e-0083000,1.200000e-002,1.333333e-0083000,1.100000e-002,1.222222e-0083000,1.200000e-002,1.333333e-0083000,1.200000e-002,1.333333e-0083000,1.200000e-002,1.333333e-0083095,1.000000e-002,1.201122e-0083202,8.000000e-003,9.499766e-0093314,2.000000e-003,2.226938e-0093425,1.000000e-002,1.105950e-0083558,8.000000e-003,8.484733e-0093701,1.000000e-002,1.011975e-0083868,1.000000e-002,9.719231e-0094000,2.000000e-002,1.250000e-0084000,2.200000e-002,1.375000e-0084000,1.000000e-002,6.250000e-0094000,1.000000e-002,6.250000e-0094000,2.000000e-002,1.250000e-0084000,1.300000e-002,8.125000e-0094000,2.000000e-002,1.250000e-0084000,2.000000e-002,1.250000e-0084000,1.300000e-002,8.125000e-0094000,1.200000e-002,7.500000e-0094025,1.000000e-002,9.375366e-0094208,1.000000e-002,8.736860e-0094413,1.000000e-002,8.361742e-0095000,1.130000e-001,4.520000e-0085000,1.120000e-001,4.480000e-0085000,1.120000e-001,4.480000e-0085000,1.120000e-001,4.480000e-0085000,6.100000e-002,2.440000e-0085000,5.300000e-002,2.120000e-0085000,5.100000e-002,2.040000e-0085000,6.100000e-002,2.440000e-0085000,6.200000e-002,2.480000e-0085000,1.120000e-001,4.480000e-0085120,7.400000e-002,1.254612e-0096931,2.200000e-002,1.595048e-0086951,1.200000e-002,1.644162e-0087435,1.200000e-002,1.537131e-0088002,1.200000e-002,1.414741e-0088661,1.200000e-002,1.307096e-0089524,2.300000e-002,3.546185e-00910000,1.320000e-001,1.320000e-00810000,6.200000e-002,6.200000e-00910000,5.900000e-002,5.900000e-00910000,5.100000e-002,5.100000e-00910000,5.100000e-002,5.100000e-00910000,1.300000e-001,1.300000e-00810000,1.320000e-001,1.320000e-00810000,1.240000e-001,1.240000e-00810000,6.100000e-002,6.100000e-00910370,5.100000e-002,6.628077e-00911264,3.170000e-001,9.993876e-01018753,4.000000e-002,9.076562e-00925032,1.330000e-001,3.889604e-00947311,1.310000e-001,5.366109e-00955515,1.710000e-001,5.292523e-00963009,1.720000e-001,5.384159e-009920683,6.895000e+000,2.895980e-009968672,1.099900e+001,2.330608e-0091081841,7.386000e+000,2.692134e-0091092610,1.011000e+001,2.159914e-009
