+++ greedy_003/types.h
