--- greedy_003/union.h
@@ -1,4 +1,0 @@
-#include <stdlib.h>
-#include "problem.h"
-
-void union_of_sets(struct set left, struct set right, struct set* unionSet);
